First step from global refactorin. Alot of static methods will be replaced by objects and it methods. Will be implemented patterns.
