+++ projects/GKCore/GDModel/Providers/GEDCOM/GEDCOMAppFormat.cs
@@ -1,0 +1,37 @@
+﻿/*
+ *  "GEDKeeper", the personal genealogical database editor.
+ *  Copyright (C) 2009-2019 by Sergey V. Zhdanovskih.
+ *
+ *  This file is part of "GEDKeeper".
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+namespace GDModel.Providers.GEDCOM
+{
+    public sealed class GEDCOMAppFormat
+    {
+        public GEDCOMFormat Format { get; private set; }
+        public string Sign { get; private set; }
+        public string Name { get; private set; }
+        public int PredefCharset { get; private set; }
+
+        public GEDCOMAppFormat(GEDCOMFormat format, string sign, string name, int predefCharset)
+        {
+            Format = format;
+            Sign = sign;
+            Name = name;
+            PredefCharset = predefCharset;
+        }
+    }
+}
